Working GoogleLogin component
